refactor(backend): extract ReportService from ReportController
ReportController had no service layer at all — rate limiting, punishment
application, both ELO deltas (REPORT_ACTIONED/REPORT_REWARDED), and
event publishing were all inline in resolve(), injecting ReportRepository
directly. The most severe layering gap found across all 6 retroactive
specs, documented in specs/006-player-reports/plan.md's Constitution
Check (Principle III, FAIL). Extracts ReportService; controller is now a
thin adapter. Preserves the exact fail-fast-before-mutation ordering
(FR-013/SC-006) and ELO-effect gating. No API/response-shape or
behavioral change.
